Update how_to.md
The description of how to format FAT16 with Mac OS X was using "Disk Utility", which does NOT format FAT16 but FAT32.
My changes describes the process with the built in Terminal.app. This comes with every Mac by default and no additional software or installation is needed to use it like in my description.